# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stock Analyser API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks/{ticker}/history:
    get:
      operationId: fetchPriceHistory
      tags: [stocks]
      summary: Get historical price data
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [1d, 1mo, 3mo, 6mo, 1y, 3y, 5y]
            default: 1y
      responses:
        "200":
          description: Historical price data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockHistory"
        "404":
          description: Ticker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /stocks/{ticker}/analysis:
    get:
      operationId: fetchStockAnalysis
      tags: [stocks]
      summary: Get full stock analysis
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full stock analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAnalysis"
        "404":
          description: Ticker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /stocks/picks:
    get:
      operationId: getStockPicks
      tags: [stocks]
      summary: Get AI-curated stock picks
      description: Fetches live data on a broad set of stocks and returns AI-selected top picks with rationale
      responses:
        "200":
          description: AI-curated stock picks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockPick"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /stocks/search:
    get:
      operationId: findStocks
      tags: [stocks]
      summary: Search for stock tickers
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSearchResult"
  /stocks/{ticker}/catalysts:
    get:
      operationId: getStockCatalysts
      tags: [stocks]
      summary: Get AI-generated bull/bear catalysts for a stock
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Catalyst analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalystResult"
        "500":
          description: Internal server error
  /stocks/{ticker}/earnings:
    get:
      operationId: getStockEarnings
      tags: [stocks]
      summary: Get next earnings date for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Earnings date result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsDateResult"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PricePoint:
      type: object
      properties:
        date:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required:
        - date
        - open
        - high
        - low
        - close
        - volume
    StockHistory:
      type: object
      properties:
        ticker:
          type: string
        prices:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"
        ma50:
          type: array
          items:
            $ref: "#/components/schemas/MAPoint"
        ma200:
          type: array
          items:
            $ref: "#/components/schemas/MAPoint"
      required:
        - ticker
        - prices
        - ma50
        - ma200
    MAPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: ["number", "null"]
      required:
        - date
        - value
    StockOverview:
      type: object
      properties:
        name:
          type: string
        ticker:
          type: string
        exchange:
          type: string
        sector:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        currency:
          type: string
        marketCap:
          type: ["number", "null"]
        employees:
          type: ["integer", "null"]
        website:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
      required:
        - name
        - ticker
        - exchange
        - currency
    PricePerformance:
      type: object
      properties:
        currentPrice:
          type: number
        previousClose:
          type: number
        change1d:
          type: ["number", "null"]
        change1dPct:
          type: ["number", "null"]
        change1w:
          type: ["number", "null"]
        change1wPct:
          type: ["number", "null"]
        change1mo:
          type: ["number", "null"]
        change1moPct:
          type: ["number", "null"]
        change6mo:
          type: ["number", "null"]
        change6moPct:
          type: ["number", "null"]
        change1y:
          type: ["number", "null"]
        change1yPct:
          type: ["number", "null"]
        week52High:
          type: ["number", "null"]
        week52Low:
          type: ["number", "null"]
        indexComparison:
          type: ["string", "null"]
        indexPerformance1y:
          type: ["number", "null"]
        outperforming:
          type: ["boolean", "null"]
      required:
        - currentPrice
        - previousClose
    QuantMetrics:
      type: object
      properties:
        pe:
          type: ["number", "null"]
        pb:
          type: ["number", "null"]
        dividendYield:
          type: ["number", "null"]
        beta:
          type: ["number", "null"]
        eps:
          type: ["number", "null"]
        forwardPE:
          type: ["number", "null"]
        priceToSales:
          type: ["number", "null"]
        debtToEquity:
          type: ["number", "null"]
        returnOnEquity:
          type: ["number", "null"]
        profitMargin:
          type: ["number", "null"]
        revenueGrowth:
          type: ["number", "null"]
        earningsGrowth:
          type: ["number", "null"]
        currentRatio:
          type: ["number", "null"]
        avgVolume:
          type: ["number", "null"]
        currentVolume:
          type: ["number", "null"]
        volumeVsAvg:
          type: ["string", "null"]
        aboveMA50:
          type: ["boolean", "null"]
        aboveMA200:
          type: ["boolean", "null"]
        ma50:
          type: ["number", "null"]
        ma200:
          type: ["number", "null"]
        trendDirection:
          type: string
          enum: [uptrend, downtrend, sideways, unknown]
        analystRating:
          type: ["string", "null"]
        targetPrice:
          type: ["number", "null"]
        targetLowPrice:
          type: ["number", "null"]
        targetHighPrice:
          type: ["number", "null"]
        targetMedianPrice:
          type: ["number", "null"]
        numberOfAnalysts:
          type: ["integer", "null"]
        forwardEps:
          type: ["number", "null"]
        forwardEps2y:
          type: ["number", "null"]
        supportLevel:
          type: ["number", "null"]
        resistanceLevel:
          type: ["number", "null"]
      required:
        - trendDirection
    GrowthAnalysis:
      type: object
      properties:
        cagr1y:
          type: ["number", "null"]
        cagr3y:
          type: ["number", "null"]
        cagr5y:
          type: ["number", "null"]
        projectedLow:
          type: ["number", "null"]
          description: "Low-case projected growth rate (%)"
        projectedHigh:
          type: ["number", "null"]
          description: "High-case projected growth rate (%)"
        realPriceLow:
          type: ["number", "null"]
          description: "Low-case projected price in actual currency (currentPrice × (1 + projectedLow/100))"
        realPriceHigh:
          type: ["number", "null"]
          description: "High-case projected price in actual currency (currentPrice × (1 + projectedHigh/100))"
        scoreLow:
          type: ["number", "null"]
          description: "Normalized score 0-100 for low case (50 + projectedLow/2)"
        scoreHigh:
          type: ["number", "null"]
          description: "Normalized score 0-100 for high case (50 + projectedHigh/2)"
        projectionBasis:
          type: string
        bullScenario:
          type: ["string", "null"]
        bearScenario:
          type: ["string", "null"]
      required:
        - projectionBasis
    SuggestionItem:
      type: object
      properties:
        label:
          type: string
        detail:
          type: string
      required:
        - label
        - detail
    Suggestion:
      type: object
      properties:
        riskProfile:
          type: string
          enum: [Aggressive, Moderate, Conservative]
        view:
          type: string
          enum: [Leaning Positive, Neutral, Leaning Negative]
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/SuggestionItem"
        investorType:
          type: array
          items:
            type: string
        disclaimer:
          type: string
      required:
        - riskProfile
        - view
        - reasons
        - investorType
        - disclaimer
    NewsItem:
      type: object
      properties:
        headline:
          type: string
        publisher:
          type: string
        url:
          type: ["string", "null"]
        publishedAt:
          type: string
        sentiment:
          type: string
          enum: [Positive, Neutral, Negative]
        impact:
          type: string
      required:
        - headline
        - publisher
        - publishedAt
        - sentiment
        - impact
    LiveNewsAnalysis:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        overallSentiment:
          type: string
          enum: [Positive, Neutral, Negative]
        sentimentScore:
          type: number
          description: "Numeric score from -5 (very negative) to +5 (very positive)"
        summary:
          type: string
        impactBullets:
          type: array
          items:
            type: string
          description: "2-4 bullet points explaining how the news affects risk, growth, or valuation"
      required:
        - items
        - overallSentiment
        - sentimentScore
        - summary
        - impactBullets
    SimpleExplanation:
      type: object
      properties:
        whatTheyDo:
          type: string
        howItsDoing:
          type: string
        whatNewsMeans:
          type: string
        whatNumbersMean:
          type: string
        whatCouldHappen:
          type: string
        whoItMightSuit:
          type: string
      required:
        - whatTheyDo
        - howItsDoing
        - whatNewsMeans
        - whatNumbersMean
        - whatCouldHappen
        - whoItMightSuit
    EarningsDateResult:
      type: object
      properties:
        earningsDate:
          type: ["string", "null"]
          description: "ISO 8601 date string of the next earnings date, or null if unknown"
      required:
        - earningsDate
    CatalystResult:
      type: object
      properties:
        ticker:
          type: string
        bulls:
          type: array
          items:
            type: string
        bears:
          type: array
          items:
            type: string
        generatedAt:
          type: string
        cached:
          type: boolean
      required:
        - ticker
        - bulls
        - bears
        - generatedAt
...
    StockAnalysis:
      type: object
      properties:
        overview:
          $ref: "#/components/schemas/StockOverview"
        performance:
          $ref: "#/components/schemas/PricePerformance"
        quant:
          $ref: "#/components/schemas/QuantMetrics"
        growth:
          $ref: "#/components/schemas/GrowthAnalysis"
        liveNews:
          $ref: "#/components/schemas/LiveNewsAnalysis"
        qualitativeNotes:
          type: ["string", "null"]
        suggestion:
          $ref: "#/components/schemas/Suggestion"
        simpleExplanation:
          $ref: "#/components/schemas/SimpleExplanation"
        generatedAt:
          type: string
        cached:
          type: boolean
          description: "Whether this response was served from cache"
        cachedAt:
          type: string
          description: "ISO timestamp of when the analysis was originally generated"
        bottomLine:
          $ref: "#/components/schemas/BottomLineResult"
      required:
        - overview
        - performance
        - quant
        - growth
        - liveNews
        - suggestion
        - simpleExplanation
        - generatedAt
        - cached
        - cachedAt
    StockPick:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        sector:
          type: ["string", "null"]
        currency:
          type: string
        currentPrice:
          type: number
        change1dPct:
          type: ["number", "null"]
        change1yPct:
          type: ["number", "null"]
        trendDirection:
          type: string
        analystRating:
          type: ["string", "null"]
        targetPrice:
          type: ["number", "null"]
        pe:
          type: ["number", "null"]
        beta:
          type: ["number", "null"]
        rationale:
          type: string
        conviction:
          type: string
          enum: [High, Medium]
        category:
          type: string
          enum: [Growth, Value, Dividend, Momentum]
      required:
        - ticker
        - name
        - currency
        - currentPrice
        - trendDirection
        - rationale
        - conviction
        - category
    StockSearchResult:
      type: object
      properties:
        ticker:
          type: string
        name:
          type: string
        exchange:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
      required:
        - ticker
        - name
...
    BottomLineScores:
      type: object
      properties:
        momentum:
          type: number
        valuation:
          type: number
        risk:
          type: number
        quality:
          type: number
      required:
        - momentum
        - valuation
        - risk
        - quality
    BottomLinePriceTargets:
      type: object
      properties:
        oneMonth:
          type: ["number", "null"]
        sixMonths:
          type: ["number", "null"]
        oneYear:
          type: ["number", "null"]
        oneMonthLow:
          type: ["number", "null"]
        oneMonthHigh:
          type: ["number", "null"]
        oneYearLow:
          type: ["number", "null"]
        oneYearHigh:
          type: ["number", "null"]
      required:
        - oneMonth
        - sixMonths
        - oneYear
        - oneMonthLow
        - oneMonthHigh
        - oneYearLow
        - oneYearHigh
    BottomLineSignals:
      type: object
      properties:
        rsi14:              { type: ["number","null"] }
        stochasticK:        { type: ["number","null"] }
        stochasticD:        { type: ["number","null"] }
        williamsR:          { type: ["number","null"] }
        macd:
          type: string
          enum: ["bullish","bearish","neutral"]
        macdLine:           { type: ["number","null"] }
        macdSignalLine:     { type: ["number","null"] }
        macdHistogram:      { type: ["number","null"] }
        bollingerUpper:     { type: ["number","null"] }
        bollingerMiddle:    { type: ["number","null"] }
        bollingerLower:     { type: ["number","null"] }
        bollingerPosition:  { type: ["number","null"] }
        atr14:              { type: ["number","null"] }
        obv:                { type: ["number","null"] }
        obvTrend:
          type: ["string","null"]
          enum: ["rising","falling","flat",null]
        roc20:              { type: ["number","null"] }
        vsMA20:             { type: ["number","null"] }
        vsMA50:             { type: ["number","null"] }
        vsMA200:            { type: ["number","null"] }
        goldenCross:        { type: ["boolean","null"] }
        ichimokuSignal:
          type: ["string","null"]
          enum: ["above_cloud","below_cloud","in_cloud",null]
        fibLevel:           { type: ["string","null"] }
        peRatio:            { type: ["number","null"] }
        forwardPE:          { type: ["number","null"] }
        pegRatio:           { type: ["number","null"] }
        pbRatio:            { type: ["number","null"] }
        psRatio:            { type: ["number","null"] }
        evToEbitda:         { type: ["number","null"] }
        earningsYield:      { type: ["number","null"] }
        dividendYield:      { type: ["number","null"] }
        grahamNumber:       { type: ["number","null"] }
        grahamVsCurrent:    { type: ["number","null"] }
        beta:               { type: ["number","null"] }
        historicalVolatility30d: { type: ["number","null"] }
        sharpeRatio:        { type: ["number","null"] }
        sortinoRatio:       { type: ["number","null"] }
        maxDrawdown:        { type: ["number","null"] }
        valueAtRisk95:      { type: ["number","null"] }
        calmarRatio:        { type: ["number","null"] }
        returnSkewness:     { type: ["number","null"] }
        returnKurtosis:     { type: ["number","null"] }
        fiftyTwoWeekPosition: { type: ["number","null"] }
        revenueGrowth:      { type: ["number","null"] }
        profitMargin:       { type: ["number","null"] }
        returnOnEquity:     { type: ["number","null"] }
        debtToEquity:       { type: ["number","null"] }
        currentRatio:       { type: ["number","null"] }
      required:
        - macd
    BottomLineResult:
      type: object
      properties:
        verdict:
          type: string
          enum: ["Strong Buy","Buy","Hold","Sell","Strong Sell"]
        verdictColour:
          type: string
          enum: ["green","amber","red"]
        compositeScore:
          type: number
        oneLiner:
          type: string
        healthStatus:
          type: string
          enum: ["Great","Good","Okay","Weak","Poor"]
        priceTargets:
          $ref: "#/components/schemas/BottomLinePriceTargets"
        targetRationale:
          type: string
        scores:
          $ref: "#/components/schemas/BottomLineScores"
        signals:
          $ref: "#/components/schemas/BottomLineSignals"
        methodsUsed:
          type: array
          items:
            type: string
        disclaimer:
          type: string
      required:
        - verdict
        - verdictColour
        - compositeScore
        - oneLiner
        - healthStatus
        - priceTargets
        - targetRationale
        - scores
        - signals
        - methodsUsed
        - disclaimer